% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpread.R
\name{findSpread}
\alias{findSpread}
\title{Find best spread}
\usage{
findSpread(p_train, v_train, k, fun, scale = TRUE)
}
\arguments{
\item{p_train}{The dataframe of training predictor dataset}

\item{v_train}{The dataframe of training response variables}

\item{k}{The numeric number of k folds}

\item{fun}{The distance function}

\item{scale}{The logic statements (TRUE/FALSE)}
}
\value{
Best spread
}
\description{
Find best spread
}
\examples{
data("met")
data("physg")
\dontrun{best.spread<-findSpread(physg,met,10,"bray",scale=TRUE)}
}
