% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpreadRdist.R
\name{findSpreadRdist}
\alias{findSpreadRdist}
\title{find best spreads using Rdist}
\usage{
findSpreadRdist(x, y, k, fun, scale = TRUE)
}
\arguments{
\item{x}{The dataframe of training predictor dataset}

\item{y}{The dataframe of training response variables}

\item{k}{The numeric number of k folds}

\item{fun}{The distance function}

\item{scale}{The logic statements (TRUE/FALSE)}
}
\value{
The vector of best spreads
}
\description{
find best spreads using Rdist
}
