% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpreadVegan.R
\name{findSpreadVegan}
\alias{findSpreadVegan}
\title{Find best spread using vegan function}
\usage{
findSpreadVegan(x, y, k, fun, scale = TRUE)
}
\arguments{
\item{x}{The dataframe of training predictor dataset}

\item{y}{The dataframe of training response variables}

\item{k}{The numeric number of k folds}

\item{fun}{The distance function}

\item{scale}{The logic statements (TRUE/FALSE)}
}
\value{
The vector of best spreads
}
\description{
Find best spread using vegan function
}
