% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnn.distance.R
\name{grnn.distance}
\alias{grnn.distance}
\title{grnn distance}
\usage{
grnn.distance(x, y, fun)
}
\arguments{
\item{x}{The matrix of training predictor dataset}

\item{y}{The matrix of training response variables}

\item{fun}{The distance function}
}
\value{
The matrix of distance between a and b
}
\description{
grnn distance
}
\examples{
data("met")
data("physg")
distance<-grnn.distance(physg[1:30,],physg[40:60,],"bray")
}
