% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\docType{data}
\name{met}
\alias{met}
\title{meteorological dataset}
\format{
A data frame with 378 rows and 11 variables:
\describe{
\item{\code{MAT}}{double COLUMN_DESCRIPTION}
\item{\code{WMMT}}{double COLUMN_DESCRIPTION}
\item{\code{CMMT}}{double COLUMN_DESCRIPTION}
\item{\code{GROWSEAS}}{double COLUMN_DESCRIPTION}
\item{\code{GSP}}{double COLUMN_DESCRIPTION}
\item{\code{MMGSP}}{double COLUMN_DESCRIPTION}
\item{\code{Three_WET}}{double COLUMN_DESCRIPTION}
\item{\code{Three_DRY}}{double COLUMN_DESCRIPTION}
\item{\code{RH}}{double COLUMN_DESCRIPTION}
\item{\code{SH}}{double COLUMN_DESCRIPTION}
\item{\code{ENTHAL}}{double COLUMN_DESCRIPTION}
}
}
\usage{
met
}
\description{
Data from a global collection by Robert A. Spicer.
It include 11 climate variables from 378 sites.
}
\details{
DETAILS
}
\keyword{datasets}
