% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physg.R
\docType{data}
\name{physg}
\alias{physg}
\title{physiognomy dataset}
\format{
A data frame with 378 rows and 31 variables:
\describe{
\item{\code{Lobed}}{double COLUMN_DESCRIPTION}
\item{\code{No.Teeth}}{double COLUMN_DESCRIPTION}
\item{\code{Regular.teeth}}{double COLUMN_DESCRIPTION}
\item{\code{Close.teeth}}{double COLUMN_DESCRIPTION}
\item{\code{Round.teeth}}{double COLUMN_DESCRIPTION}
\item{\code{Acute.teeth}}{double COLUMN_DESCRIPTION}
\item{\code{Compound.teeth}}{double COLUMN_DESCRIPTION}
\item{\code{Nanophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Leptophyll.1}}{double COLUMN_DESCRIPTION}
\item{\code{Leptophyll.2}}{double COLUMN_DESCRIPTION}
\item{\code{Microphyll.1}}{double COLUMN_DESCRIPTION}
\item{\code{Microphyll.2}}{double COLUMN_DESCRIPTION}
\item{\code{Microphyll.3}}{double COLUMN_DESCRIPTION}
\item{\code{Mesophyll.1}}{double COLUMN_DESCRIPTION}
\item{\code{Mesophyll.2}}{double COLUMN_DESCRIPTION}
\item{\code{Mesophyll.3}}{double COLUMN_DESCRIPTION}
\item{\code{Emarginate.apex}}{double COLUMN_DESCRIPTION}
\item{\code{Round.apex}}{double COLUMN_DESCRIPTION}
\item{\code{Acute.apex}}{double COLUMN_DESCRIPTION}
\item{\code{Attenuate.apex}}{double COLUMN_DESCRIPTION}
\item{\code{Cordate.base}}{double COLUMN_DESCRIPTION}
\item{\code{Round.base}}{double COLUMN_DESCRIPTION}
\item{\code{Acute.base}}{double COLUMN_DESCRIPTION}
\item{\code{L.W..1.1}}{double COLUMN_DESCRIPTION}
\item{\code{L.W.1.2.1}}{double COLUMN_DESCRIPTION}
\item{\code{L.W.2.3.1}}{double COLUMN_DESCRIPTION}
\item{\code{L.W.3.4.1}}{double COLUMN_DESCRIPTION}
\item{\code{L.W..4.1}}{double COLUMN_DESCRIPTION}
\item{\code{Obovate}}{double COLUMN_DESCRIPTION}
\item{\code{Elliptic}}{double COLUMN_DESCRIPTION}
\item{\code{Ovate}}{double COLUMN_DESCRIPTION}
}
}
\usage{
physg
}
\description{
Data from a global collection by Robert A. Spicer.
It include 31 leaf physiognomies variables from 378 sites.
}
\details{
DETAILS
}
\keyword{datasets}
