% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg.distance.R
\name{veg.distance}
\alias{veg.distance}
\title{distance using vegdist}
\usage{
veg.distance(a, b, fun = "bray")
}
\arguments{
\item{a}{The matrix of training predictor dataset}

\item{b}{The matrix of validation predictor dataset}

\item{fun}{The distance function}
}
\value{
The matrix of distance between a and b
}
\description{
distance using vegdist
}
\examples{
data("physg")
distance<-veg.distance(physg[1:10,],physg[20:40,],"bray")
}
