% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnn.distance.R
\name{grnn.distance}
\alias{grnn.distance}
\title{grnn distance}
\usage{
grnn.distance(x, y, fun)
}
\arguments{
\item{x}{The dataframe of training predictor dataset}

\item{y}{The dataframe of training response variables}

\item{fun}{The distance function}
}
\value{
The matrix of distance between a and b
}
\description{
grnn distance
}
\examples{
data("physg")
physg.train<-physg[1:10,]
physg.test<-physg[11:30,]
distance<-grnn.distance(physg.test,physg.train,"bray")
}
