% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg.distance.R
\name{veg.distance}
\alias{veg.distance}
\title{distance using vegdist}
\usage{
veg.distance(a, b, fun = "bray")
}
\arguments{
\item{a}{The dataframe of training predictor dataset}

\item{b}{The dataframe of validation predictor dataset}

\item{fun}{The distance function}
}
\value{
The matrix of distance between a and b
}
\description{
distance using vegdist
}
\examples{
data("physg")
physg.train<-physg[1:10,]
physg.test<-physg[11:30,]
distance<-veg.distance(physg.test,physg.train,"bray")
}
